abeci-pangrams works
